Split out completer into separate module
